src/web/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-jazzmin==3.0.0
django-mptt
djangorestframework==3.15.1
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.5.3
pydantic_core==2.14.6
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.9.0
django-phonenumber-field[phonenumbers]==7.3.0
yarl==1.9.4
psycopg2-binary==2.9.9